Add files via upload
This fork enhances V1.1 with improved arithmetic precision and easier configuration, including a pretty print of the current config.

V2.1 has one incompatibility with V1 -- the #define SHUNT_RESISTOR_VALUE units are changed to from ohms to milli-ohms.  The Arduino IDE with default compiler warnings will flag an existing #define in your sketch so you can adjust it (or remove it).